+++ src/ts/util/editorCommenEvent.ts
@@ -23,14 +23,20 @@ export const focusEvent = (vditor: IVditor, editorElement: HTMLElement) => {
     });
 };
 
-export const scrollCenter = (editorElement: HTMLElement) => {
-    const cursorTop = getCursorPosition(editorElement).top;
-    const center = editorElement.clientHeight / 2;
-    if (cursorTop > center) {
-        editorElement.scrollTop = editorElement.scrollTop + (cursorTop - center);
+export const scrollCenter = (vditor: IVditor) => {
+    if (vditor.options.typewriterMode && typeof vditor.options.height === "string") {
+        window.scrollTo(window.scrollX, vditor.element.clientHeight + vditor.element.offsetTop - window.innerHeight);
     }
-    if (cursorTop < 0) {
-        editorElement.scrollTop = editorElement.scrollTop + cursorTop;
+    if (typeof vditor.options.height === "number") {
+        const editorElement = vditor[vditor.currentMode].element;
+        const cursorTop = getCursorPosition(editorElement).top;
+        const center = editorElement.clientHeight / 2;
+        if (cursorTop > center) {
+            editorElement.scrollTop = editorElement.scrollTop + (cursorTop - center);
+        }
+        if (cursorTop < 0) {
+            editorElement.scrollTop = editorElement.scrollTop + cursorTop;
+        }
     }
 };