+++ CHANGELOG.md
@@ -59,11 +59,15 @@
 
 ### v3.1.4 / 2020-04-0x
 
+* [290](https://github.com/Vanessa219/vditor/pull/290) 🎨 add minHeight `改进功能`
+
 ### v3.1.3 / 2020-04-08
 
 * [71](https://github.com/Vanessa219/vditor/issues/71) 优化移动端体验 `改进功能`
 * [283](https://github.com/Vanessa219/vditor/issues/283) 添加 SetSetext 配置 `引入特性`
 * [278](https://github.com/Vanessa219/vditor/issues/278) IR 细节修改 `修复缺陷`
+* 文档更新
+  * 添加 `options.minHeight`
 
 ### v3.0.12 / 2020-04-06